requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
opencv-python
numpy
pandas
pillow
